# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: louyi-admin 是一個輕量級的 Java 快速開發平臺，能快速開發項目並交付【接私活利器】
  version: 1.0.0
  title: 樓易健康科技管理系統
basePath: /louyi-admin
schemes:
  - http

#認證
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定義介面數據
paths:
...
  /sys/user/list:
    get:
      tags:
        - 管理員管理
      summary: 管理員列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序欄位
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 用戶名
          in: query
          type: string
      responses:
        '200':
          description: 返回管理員列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 管理員管理
      summary: 獲取管理員信息
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用戶ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回管理員信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 管理員管理
      summary: 修改密碼
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - 管理員管理
      summary: 添加管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - 管理員管理
      summary: 修改管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - 管理員管理
      summary: 刪除管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 用戶ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/list:
    get:
      tags:
        - 角色管理
      summary: 角色列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序欄位
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: 角色名
          in: query
          type: string
      responses:
        '200':
          description: 返回角色列表
          schema:
            $ref: '#/definitions/SysRoleEntityList'
...
  /sys/role/info/{roleId}:
    get:
      tags:
        - 角色管理
      summary: 獲取角色信息
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              role:
                $ref: '#/definitions/SysRoleEntity'
  /sys/role/save:
    post:
      tags:
        - 角色管理
      summary: 添加角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/update:
    post:
      tags:
        - 角色管理
      summary: 修改角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/delete:
    post:
      tags:
        - 角色管理
      summary: 刪除角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      msg:
        description: 失敗原因
        type: string
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 原密碼
        type: string
      newPassword:
        description: 新密碼
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 用戶ID
        type: integer
        format: int64
      username:
        description: 用戶名
        type: string
      password:
        description: 密碼
        type: string
      email:
        description: 郵箱
        type: string
      mobile:
        description: 手機號
        type: string
      status:
        description: 狀態  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 創建者ID
        type: integer
        format: int64
      createTime:
        description: 創建時間
        type: string
        format: date-time
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
    type: object
    properties:
      userId:
        description: 用戶ID
        type: integer
        format: int64
      username:
        description: 用戶名
        type: string
      password:
        description: 密碼
        type: string
      email:
        description: 郵箱
        type: string
      mobile:
        description: 手機號
        type: string
      status:
        description: 狀態  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名稱
        type: string
      remark:
        description: 備註
        type: string
      menuIdList:
        description: 菜單ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 創建者ID
        type: integer
        format: int64
      createTime:
        description: 創建時間
        type: string
        format: date-time
  SysRoleEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysRoleEntity'
  SysRoleEntityEdit:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名稱
        type: string
      remark:
        description: 備註
        type: string
      menuIdList:
        description: 菜單ID列表
        type: array
        items:
          type: integer
          format: int64
...